// generated/src/app/api-models/user-defined-valdt-codes.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class UserDefinedValdtCodes {
  validatedTableName: string;
  validatedColumnName: string;
  userValidCode: string;
  languageId: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  userValidCodeShortDesc: string;
  userValidCodeLongDesc: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}